% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer of the selected state}

\item{year}{a string of the selected year}
}
\value{
a plot of the map
}
\description{
Make a mapping of the accidents of the selected state and selected year
}
\examples{
\dontrun{fars_map_state(1,"2013")}
}
