% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Importing a data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string that shows the pathname of the file}
}
\value{
a dataframe containing the data of the filename
}
\description{
Importing data from a file name
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
}
