% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creating a list of dataframes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a list of selected years}
}
\value{
a list of the related dataframes
}
\description{
Create a list of dataframes of selected years
}
\examples{
\dontrun{fars_read_years(c("2013","2014"))}
}
