% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Make a summary of the number of accident per year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a list of selected years}
}
\value{
a tibble summing up the number of accidents per selected year
}
\description{
Make tibble summing up the number of accidents per selected year
}
\examples{
\dontrun{make_filename("2013")}
}
