% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a string of the selected year}
}
\value{
a string of the related filename
}
\description{
Make a filename of the select year
}
\examples{
\dontrun{make_filename("2013")}
}
